Move span annotation components to common components folder
